+++ cmd/pint/lint.go
@@ -28,7 +28,10 @@ func actionLint(c *cli.Context) (err error) {
 	if err != nil {
 		return fmt.Errorf("failed to load config file %q: %w", c.Path(configFlag), err)
 	}
-	cfg.SetDisabledChecks(c.Bool(offlineFlag), c.StringSlice(disabledFlag))
+	cfg.SetDisabledChecks(c.StringSlice(disabledFlag))
+	if c.Bool(offlineFlag) {
+		cfg.DisableOnlineChecks()
+	}
 
 	d := discovery.NewGlobFileFinder()
 	toScan, err := d.Find(paths...)